requirements.txt
Flask
pillow
requests
beautifulsoup4
pandas
openpyxl
